1、修正点击搜索结果后无法显示详情的BUG以及切换游戏版本搜索不到结果的BUG

+++ 饥荒百科全书CSharp/Class/Global.cs
@@ -61,7 +61,7 @@ namespace 饥荒百科全书CSharp.Class
         /// <summary>
         /// 游戏版本
         /// </summary>
-        public static int GameVersion { get; set; }
+        public static double GameVersion { get; set; }
 
         /// <summary>
         /// 内置游戏版本Json文件夹名
